+++ Assets/PatchKit Patcher/Scripts/AppUpdater/AppUpdaterDiffStrategy.cs
@@ -1,11 +1,12 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using PatchKit.Unity.Patcher.AppUpdater.Commands;
 using PatchKit.Unity.Patcher.Cancellation;
 using PatchKit.Unity.Patcher.Debug;
 
 namespace PatchKit.Unity.Patcher.AppUpdater
 {
-    public class AppUpdaterDiffStrategy : IAppUpdaterStrategy
+    public class AppUpdaterDiffStrategy: IAppUpdaterStrategy
     {
         private struct DiffCommands
         {
@@ -28,6 +29,11 @@ namespace PatchKit.Unity.Patcher.AppUpdater
             _context = context;
         }
 
+        public StrategyType GetStrategyType()
+        {
+            return StrategyType.Diff;
+        }
+
         public void Update(CancellationToken cancellationToken)
         {
             Assert.MethodCalledOnlyOnce(ref _updateHasBeenCalled, "Update");
@@ -43,7 +49,7 @@ namespace PatchKit.Unity.Patcher.AppUpdater
 
             var commandFactory = new AppUpdaterCommandFactory();
             var geolocateCommand = commandFactory.CreateGeolocateCommand();
-            
+
             geolocateCommand.Prepare(_context.StatusMonitor);
             geolocateCommand.Execute(cancellationToken);