+++ Assets/PatchKit Patcher/Scripts/AppUpdater/AppUpdaterStrategyResolver.cs
@@ -1,11 +1,12 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using PatchKit.Unity.Patcher.AppUpdater.Commands;
 using PatchKit.Unity.Patcher.Cancellation;
 using PatchKit.Unity.Patcher.Debug;
 
 namespace PatchKit.Unity.Patcher.AppUpdater
 {
-    public class AppUpdaterStrategyResolver : IAppUpdaterStrategyResolver
+    public class AppUpdaterStrategyResolver: IAppUpdaterStrategyResolver
     {
         private static readonly DebugLogger DebugLogger = new DebugLogger(typeof(AppUpdaterStrategyResolver));
 
@@ -14,7 +15,37 @@ namespace PatchKit.Unity.Patcher.AppUpdater
             DebugLogger.LogConstructor();
         }
 
-        public IAppUpdaterStrategy Resolve(AppUpdaterContext context)
+        public IAppUpdaterStrategy Create(StrategyType type, AppUpdaterContext context)
+        {
+            switch (type)
+            {
+                case StrategyType.Empty:
+                    return new AppUpdaterEmptyStrategy();
+                case StrategyType.Content:
+                    return new AppUpdaterContentStrategy(context);
+                case StrategyType.Diff:
+                    return new AppUpdaterDiffStrategy(context);
+                default:
+                    return new AppUpdaterContentStrategy(context);
+            }
+        }
+
+        public StrategyType GetFallbackStrategy(StrategyType strategyType)
+        {
+            switch (strategyType)
+            {
+                case StrategyType.Empty:
+                    return StrategyType.Empty;
+                case StrategyType.Content:
+                    return StrategyType.None;
+                case StrategyType.Diff:
+                    return StrategyType.Content;
+                default:
+                    return StrategyType.Content;
+            }
+        }
+
+        public StrategyType Resolve(AppUpdaterContext context)
         {
             Checks.ArgumentNotNull(context, "context");
 
@@ -29,7 +60,7 @@ namespace PatchKit.Unity.Patcher.AppUpdater
                 {
                     DebugLogger.Log("Installed version is the same as the latest version. Using empty strategy.");
 
-                    return new AppUpdaterEmptyStrategy();
+                    return StrategyType.Empty;
                 }
 
                 if (installedVersionId < latestVersionId)
@@ -39,7 +70,7 @@ namespace PatchKit.Unity.Patcher.AppUpdater
                         "Installed version is older than the latest version. " +
                         "Using content strategy due to define PATCHKIT_DONT_USE_DIFF_UPDATES");
                     
-                    return new AppUpdaterContentStrategy(context);
+                    return StrategyType.Content;
 #else
                     DebugLogger.Log(
                         "Installed version is older than the latest version. Checking whether cost of updating with diff is lower than cost of updating with content...");
@@ -75,7 +106,7 @@ namespace PatchKit.Unity.Patcher.AppUpdater
                             DebugLogger.Log(
                                 "Version is not consistent. Diff update is forbidden - using content strategy.");
 
-                            return new AppUpdaterContentStrategy(context);
+                            return StrategyType.Content;
                         }
                     }
 
@@ -94,7 +125,7 @@ namespace PatchKit.Unity.Patcher.AppUpdater
                         DebugLogger.Log(
                             "Cost of updating with diff is lower than cost of updating with content. Using diff strategy.");
 
-                        return new AppUpdaterDiffStrategy(context);
+                        return StrategyType.Diff;
                     }
 
                     DebugLogger.Log(
@@ -110,8 +141,8 @@ namespace PatchKit.Unity.Patcher.AppUpdater
             {
                 DebugLogger.Log("Application is not installed. Using content strategy.");
             }
-            
-            return new AppUpdaterContentStrategy(context);
+
+            return StrategyType.Content;
         }
 
         private ulong GetContentCost(AppUpdaterContext context)